+++ CHANGELOG.md
@@ -8,7 +8,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 
 ### Added
-
+- Printing coverage report on the bottom of the pdf case report 
 - Logs uri without pwd when connecting
 - Disease-causing transcripts in case report
 - Thicker lines in case report